perf: complete operator runtime refactor
Collapse the operator-runtime performance work into one coherent change.

This wires semantic benchmark gates and local CI coverage, moves the hot scan/join/aggregate/sort paths onto typed runtime structures, hardens spill and reclaimer ownership, and retires legacy compatibility seams around row formats, specs, lowering, and scheduler work units.

It also locks in the review follow-up guardrails for aggregate DISTINCT/ORDERED replay, classic IEJoin cursor emission, runtime filters, profile output, and operator-runtime naming.

+++ crates/storage/src/row/partition/partitioned.rs
@@ -213,7 +213,7 @@ mod tests {
         if store.is_empty() {
             return Vec::new();
         }
-        let pinned = store.pin_ordinal_range(0, store.count() as u32).unwrap();
+        let pinned = store.pin_ordinal_range(0, store.count()).unwrap();
         let mut out = test_chunk_with_capacity(&[LogicalType::Integer], store.count() as usize);
         pinned.gather_columns(&[0], &mut out, 0).unwrap();
         (0..store.count() as usize)
